#12 cleanup
