+++ .gitignore
@@ -50,3 +50,5 @@ next-env.d.ts
 
 # archive
 /src/archive
+
+/src/experimental
