+++ package.json
@@ -17,6 +17,7 @@
     "@llamaindex/core": "^0.6.19",
     "@llamaindex/deepseek": "^0.0.31",
     "@llamaindex/huggingface": "^0.1.24",
+    "@llamaindex/workflow": "^1.1.20",
     "@llamaindex/workflow-core": "^1.2.0",
     "@next/mdx": "^15.3.3",
     "@radix-ui/react-accordion": "^1.2.11",
